Remove the deprecated otelgrpc.StreamClientInterceptor
Fix #7105
